+++ Adapter/PlentymarketsAdapter/ResponseParser/Order/OrderResponseParser.php
@@ -349,6 +349,14 @@ class OrderResponseParser implements OrderResponseParserInterface
             return null;
         }
 
+        if (null === $entry['customerData']['gender']) {
+            $this->logger->info('customer gender not defined', [
+                'customerNumber' => $entry['customerData']['number'],
+            ]);
+
+            return null;
+        }
+
         $customer = new Customer();
         $customer->setType(Customer::TYPE_NORMAL);
         $customer->setNumber($entry['customerData']['number']);
@@ -674,20 +682,27 @@ class OrderResponseParser implements OrderResponseParserInterface
             return $date['typeId'] === 8;
         });
 
+        if (!empty($shippingDate)) {
+            $shippingDate = array_shift($shippingDate);
+
+            $shippingDate = DateTimeImmutable::createFromFormat(
+                DATE_ATOM,
+                $shippingDate['date']
+            );
+        } else {
+            $now = new \DateTime();
+            $shippingDate = DateTimeImmutable::createFromFormat(
+                DATE_ATOM,
+                $now->format(DATE_ATOM)
+            );
+        }
+
         $result = [];
         foreach ($numbers as $number) {
             $package = new Package();
             $package->setShippingCode((string) $number);
             $package->setShippingProvider();
-
-            if (!empty($shippingDate)) {
-                $shippingDate = array_shift($shippingDate);
-
-                $package->setShippingTime(DateTimeImmutable::createFromFormat(
-                    DATE_ATOM,
-                    $shippingDate['date']
-                ));
-            }
+            $package->setShippingTime($shippingDate);
 
             $result[] = $package;
         }
@@ -706,6 +721,14 @@ class OrderResponseParser implements OrderResponseParserInterface
 
         foreach ($entry['orderItems'] as $item) {
             $number = $this->getNumberFromVariation($item['itemVariationId']);
+
+            if (empty($number)) {
+                $this->logger->info('product not found', [
+                    'itemVariationId' => $item['itemVariationId'],
+                ]);
+
+                continue;
+            }
 
             $orderItem = new OrderItem();
             $orderItem->setQuantity((float) $item['quantity']);
@@ -737,7 +760,7 @@ class OrderResponseParser implements OrderResponseParserInterface
     }
 
     /**
-     * @param int $variationId
+     * @param $variationId
      *
      * @return string
      */
@@ -749,9 +772,7 @@ class OrderResponseParser implements OrderResponseParserInterface
             $response = $this->client->request('GET', 'items/variations', ['id' => $variationId]);
 
             if (empty($response)) {
-                $variations[$variationId] = null;
-
-                return $variations[$variationId];
+                return '';
             }
 
             $variation = array_shift($response);